avoid running unsupported setDefaultColor in tests for older back
