+++ .changeset/moody-parrots-brake.md
@@ -1,0 +1,6 @@
+---
+"@milaboratories/pl-client": minor
+"@milaboratories/pl-tree": minor
+---
+
+Support auth:v2 authentication flow
