+++ lib/node/pl-client/src/core/client.ts
@@ -47,19 +47,6 @@ function alternativeRootFieldName(alternativeRoot: string): string {
   return `alternative_root_${alternativeRoot}`;
 }
 
-// Parses leading "<major>.<minor>.<patch>" from a version string like
-// "3.1.1" or "3.1.1-rc1" and returns true if the parsed version is >= target.
-// Returns false for unparseable versions (safer to assume an old backend).
-function isVersionAtLeast(version: string, target: [number, number, number]): boolean {
-  const match = /^v?(\d+)\.(\d+)\.(\d+)/.exec(version);
-  if (!match) return false;
-  const parsed: [number, number, number] = [Number(match[1]), Number(match[2]), Number(match[3])];
-  for (let i = 0; i < 3; i++) {
-    if (parsed[i] !== target[i]) return parsed[i] > target[i];
-  }
-  return true;
-}
-
 /** Client to access core PL API. */
 export class PlClient {
   private readonly drivers = new Map<string, PlDriver>();
@@ -89,9 +76,6 @@ export class PlClient {
 
   /** Stores client root (this abstraction is intended for future implementation of the security model) */
   private _clientRoot: OptionalSignedResourceId = NullSignedResourceId;
-
-  // method setDefaultColor is safe to use in transactions
-  private _supportsSetDefaultColor: boolean = false;
 
   private _userResources?: UserResources;
 
@@ -279,8 +263,6 @@ export class PlClient {
       this._ll = await this.buildLLPlClient(true, wireProtocol);
     }
 
-    this.detectBackendCompatibility(info);
-
     const userRoot = await this.userResources.getUserRoot({ createIfNotExists: true });
 
     if (this.conf.alternativeRoot === undefined) {
@@ -301,10 +283,6 @@ export class PlClient {
         return await altRoot.globalId;
       });
     }
-  }
-
-  private detectBackendCompatibility(serverInfo: MaintenanceAPI_Ping_Response): void {
-    this._supportsSetDefaultColor = isVersionAtLeast(serverInfo.coreVersion, [3, 3, 0]);
   }
 
   /** Returns true if field existed */
@@ -352,8 +330,7 @@ export class PlClient {
 
         // Auto-set default color proof from the client root's signature.
         // Skip when backend doesn't implement the `set_default_color` TX request.
-        // See `detectBackendCompatibility`.
-        if (this._supportsSetDefaultColor && !isNullSignedResourceId(clientRoot) && writable) {
+        if (this.ll.supportsSetDefaultColor && !isNullSignedResourceId(clientRoot) && writable) {
           const parsed = parseSignedResourceId(clientRoot);
           if (parsed.signature) {
             tx.setDefaultColor(parsed.signature as ColorProof);
